pyproject.toml
[project]
name = "rainbow-rag"
version = "0.1.0"
description = "Simple RAG for Rainbow Bazaar PDF"
requires-python = ">=3.10"
dependencies = [
    "langchain==0.1.14",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-text-splitters>=0.0.1",
    "streamlit>=1.31.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "pypdf>=3.17.0",
    "faiss-cpu>=1.7.4",
]
